extending custom test cases

+++ utils/src/main/java/com/epam/utils/StringUtils.java
@@ -1,7 +1,8 @@
 package com.epam.utils;
 
+
 public class StringUtils {
     public static boolean isPositiveNumber(String str) {
-        return Long.parseLong(str) > 0l;
+        return Double.parseDouble(str) > 0d;
     }
 }
